Move FakeClient to common package and adapt Generic Type
